Fix /selftimeout parameter handling (#14)
Problem: when using `/selftimeout`, if you *only* specify the duration
in hours, then an unexpected extra hour is added to the time.

Result: pain and suffering. I have to wait an extra hour before ~~being
offtopic in general~~ participating in productive conversations.

Solution: match on the `hours` value when minutes isn't specified. If
it's `0`, then the current default (60) is still used. If it's nonzero,
use 0 instead.