feat: remove /modmail_setup command
It doesn't work as intended since if you delete the modmail message, the previous deleted message ID is still stored in the data struct, and so the command thinks it's still valid and does nothing.

The way to setup the modmail message it is to delete the modmail message and restart the bot, because it automatically configures the modmail on startup.