Default to 2024 edition (#58)
Since 2024 edition was just released, we can now default to it.

I believe I got everything, but please do a double check just to make
sure. Let me know if there are any other thoughts.

Also, one note that as of this current comment date, godbolt has not
_yet_ upgraded to 1.85 (I'm sure they will soon), so technically godbolt
will be broken for a short time if this is merged before they update

---------

Signed-off-by: Cherry <13651622+MolotovCherry@users.noreply.github.com>