Feat: Add support for Tree borrows with miri (#67)
Mostly self explanatory change. I think this should work, but I haven't
tested the bot or anything locally. I'm not sure if you'd prefer a
different way to do the help message when only one command takes a flag,
but this seemed to be most consistent.

Also, I wasn't sure if Ferris should take in the flag as
`aliasing_model` or `aliasingModel`. This would be the first two word
flag, so not sure if you prefer one way or another. I'd be happy to
change to camel case if you want. Thanks!

---------

Co-authored-by: Ivan Borgia Dardi <ivanbdardi@gmail.com>