Fix: Miri Tree Borrows aliasingModel camel cased and appease clippy (#68)
Poise doesn't allow keys to contain special characters without enclosing
them in quotes, so this changes `aliasing_model` to be camelCased (also
makes it much easier to use on mobile!). I tested this locally, so it
should work now.

I also fixed Clippy in separate commits. The `format_push_string` lint
is a bit contentious since it requires handling an error that should
never occur. I fixed that in a separate commit, so if you'd prefer not
doing that I can revert the commit and disable that lint.